+++ Bundle/CoreBundle/src/DataTable/Action/LinkActionType.php
@@ -1,0 +1,110 @@
+<?php
+
+namespace Umbrella\CoreBundle\DataTable\Action;
+
+use Symfony\Component\OptionsResolver\Options;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Routing\RouterInterface;
+use Twig\Environment;
+
+class LinkActionType extends ActionType
+{
+    protected RouterInterface $router;
+
+    public function __construct(RouterInterface $router)
+    {
+        $this->router = $router;
+    }
+
+    public function render(Environment $twig, array $options): string
+    {
+        $vars = [];
+
+        $url = $options['route'] ? $this->router->generate($options['route'], $options['route_params']) : (string) $options['url'];
+
+        if ($options['xhr']) {
+            $vars['attr']['href'] = '';
+            $vars['attr']['data-xhr'] = $url;
+
+            if ($options['spinner']) {
+                $vars['attr']['data-spinner'] = 'true';
+            }
+
+            if (!empty($options['confirm'])) {
+                $vars['attr']['data-confirm'] = $options['confirm'];
+            }
+        } else {
+            $vars['attr']['href'] = $url;
+            $vars['attr']['target'] = $options['target'];
+        }
+
+        if ($options['title']) {
+            $vars['attr']['href'] = $url;
+        }
+
+        if ($options['title']) {
+            $vars['attr']['title'] = $options['title'];
+            $vars['attr']['data-bs-toggle'] = 'tooltip';
+        }
+
+        $vars['attr']['class'] = $options['class'];
+
+        $vars['icon'] = $options['icon'];
+        $vars['text'] = $options['text'];
+        $vars['translation_domain'] = $options['translation_domain'];
+
+        return $twig->render('@UmbrellaCore/DataTable/Action/link.html.twig', $vars);
+    }
+
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver
+            ->setDefault('class', null)
+            ->setAllowedTypes('class', ['null', 'string']);
+
+        $resolver
+            ->setDefault('title', null)
+            ->setAllowedTypes('title', ['null', 'string']);
+
+        $resolver
+            ->setDefault('icon', null)
+            ->setAllowedTypes('icon', ['null', 'string']);
+
+        $resolver
+            ->setDefault('text', null)
+            ->setAllowedTypes('text', ['null', 'string']);
+
+        $resolver
+            ->setDefault('translation_domain', null)
+            ->setAllowedTypes('translation_domain', ['null', 'string', 'bool'])
+            ->setNormalizer('translation_domain', fn (Options $options, $value) => true === $value ? null : $value);
+
+        $resolver
+            ->setDefault('route', null)
+            ->setAllowedTypes('route', ['null', 'string']);
+
+        $resolver
+            ->setDefault('route_params', [])
+            ->setAllowedTypes('route_params', 'array');
+
+        $resolver
+            ->setDefault('url', null)
+            ->setAllowedTypes('url', ['null', 'string']);
+
+        $resolver
+            ->setDefault('target', null)
+            ->setAllowedValues('target', [null, '_blank', '_self']);
+
+        $resolver
+            ->setDefault('xhr', false)
+            ->setAllowedTypes('xhr', 'bool');
+
+        $resolver
+            ->setDefault('confirm', null)
+            ->setAllowedTypes('confirm', ['null', 'string']);
+
+        $resolver
+            ->setDefault('spinner', false)
+            ->setAllowedTypes('spinner', 'bool');
+    }
+}